Use the new component for the Samples page
