+++ frontend/src/App.tsx
@@ -1,10 +1,9 @@
 import { Routes, Route } from "react-router-dom";
 import RequestsPage from "./pages/requests/RequestsPage";
 import PatientsPage from "./pages/patients/PatientsPage";
-import SamplesPage from "./pages/samples/SamplesPage";
 import CohortsPage from "./pages/cohorts/CohortsPage";
 import { LoginSuccessPage } from "./pages/auth/LoginSuccessPage";
-import { SamplesPage2 } from "./pages/samples/SamplesPage2";
+import { SamplesPage } from "./pages/samples/SamplesPage";
 import { NavBar } from "./shared/components/NavBar";
 import { BrowserRouter } from "react-router-dom";
 import { Providers } from "./components/Providers";
@@ -23,8 +22,6 @@ export default function App() {
             <Route path="/requests/" element={<RequestsPage />}>
               <Route path=":igoRequestId" />
             </Route>
-            {/** TODO: Remove this after testing */}
-            <Route path="/test/" element={<SamplesPage2 />} />
             <Route path="/patients/" element={<PatientsPage />}>
               <Route path=":patientId" />
             </Route>
