+++ frontend/src/shared/components/NavBar.tsx
@@ -38,9 +38,6 @@ export function NavBar() {
           <Link className="nav-link" to="/cohorts">
             Cohorts
           </Link>
-          <Link className="nav-link" to="/test">
-            Test
-          </Link>
         </Nav>
         {userEmail && (
           <div className="ms-auto d-none d-md-flex">
